ieee1212-config-rom: entry: add module for directory entry
This commit adds entry module to split structure and enumeration from
toplevel module.
