+++ libs/ieee1212-config-rom/src/lib.rs
@@ -1,67 +1,8 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // Copyright (c) 2020 Takashi Sakamoto
-#[derive(Debug)]
-pub struct Entry<'a> {
-    pub key: KeyType,
-    pub data: EntryData<'a>,
-}
+pub mod entry;
 
-#[derive(Debug)]
-pub enum EntryData<'a> {
-    Immediate(u32),
-    CsrOffset(usize),
-    Leaf(&'a [u8]),
-    Directory(Vec<Entry<'a>>),
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum KeyType {
-    Descriptor,
-    BusDependentInfo,
-    Vendor,
-    HardwareVersion,
-    Module,
-    NodeCapabilities,
-    Eui64,
-    Unit,
-    SpecifierId,
-    Version,
-    DependentInfo,
-    UnitLocation,
-    Model,
-    Instance,
-    Keyword,
-    Feature,
-    ModifiableDescriptor,
-    DirectoryId,
-    Reserved(u8),
-}
-
-impl From<u8> for KeyType {
-    fn from(val: u8) -> Self {
-        match val {
-            0x01 => KeyType::Descriptor,
-            0x02 => KeyType::BusDependentInfo,
-            0x03 => KeyType::Vendor,
-            0x04 => KeyType::HardwareVersion,
-            0x07 => KeyType::Module,
-            0x0c => KeyType::NodeCapabilities,
-            0x0d => KeyType::Eui64,
-            0x11 => KeyType::Unit,
-            0x12 => KeyType::SpecifierId,
-            0x13 => KeyType::Version,
-            0x14 => KeyType::DependentInfo,
-            0x15 => KeyType::UnitLocation,
-            0x17 => KeyType::Model,
-            0x18 => KeyType::Instance,
-            0x19 => KeyType::Keyword,
-            0x1a => KeyType::Feature,
-            0x1f => KeyType::ModifiableDescriptor,
-            0x20 => KeyType::DirectoryId,
-            _ => KeyType::Reserved(val),
-        }
-    }
-}
+use entry::*;
 
 pub fn get_root_entry_list<'a>(mut data: &'a [u8]) -> Vec<Entry<'a>> {
     // Get block of bus information.
